requirements.txt
langchain>=0.1.0
langchain-ollama>=0.1.0
langgraph>=0.0.20
tweepy>=4.14.0
python-dotenv>=1.0.0
requests>=2.31.0

